--- rules-tests/CodingStyle/Rector/FunctionLike/FuncLikeToFirstClassCallableRector/Fixture/skip_byref_usage.php.inc
@@ -1,10 +1,0 @@
-<?php
-
-namespace Rector\Tests\CodingStyle\Rector\FunctionLike\FuncLikeToFirstClassCallableRector\Fixture;
-
-function (&$foo)
-{
-    return Foo::foo($foo);
-};
-
-?>
