--- rules-tests/CodingStyle/Rector/FunctionLike/FuncLikeToFirstClassCallableRector/Fixture/skip_functions_without_calls.php.inc
@@ -1,12 +1,0 @@
-<?php
-
-namespace Rector\Tests\CodingStyle\Rector\FunctionLike\FuncLikeToFirstClassCallableRector\Fixture;
-
-function ($foo)
-{
-    return $foo;
-};
-
-fn ($foo) => $foo;
-
-?>
