--- rules-tests/CodingStyle/Rector/FunctionLike/FuncLikeToFirstClassCallableRector/Fixture/supports_varadic_and_unpack.php.inc
@@ -1,22 +1,0 @@
-<?php
-
-namespace Rector\Tests\CodingStyle\Rector\FunctionLike\FuncLikeToFirstClassCallableRector\Fixture;
-
-function (...$foo)
-{
-    return FooBar::foo(...$foo);
-};
-
-fn (...$foo) => FooBar::foo(...$foo);
-
-?>
------
-<?php
-
-namespace Rector\Tests\CodingStyle\Rector\FunctionLike\FuncLikeToFirstClassCallableRector\Fixture;
-
-FooBar::foo(...);
-
-FooBar::foo(...);
-
-?>
