--- rules-tests/CodingStyle/Rector/FunctionLike/FuncLikeToFirstClassCallableRector/config/configured_rule.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-use Rector\CodingStyle\Rector\FunctionLike\FuncLikeToFirstClassCallableRector;
-use Rector\Config\RectorConfig;
-
-return RectorConfig::configure()
-    ->withRules([FuncLikeToFirstClassCallableRector::class]);
