+++ rules-tests/CodingStyle/Rector/FunctionLike/FunctionLikeToFirstClassCallableRector/Fixture/skip_byref_usage.php.inc
@@ -1,0 +1,10 @@
+<?php
+
+namespace Rector\Tests\CodingStyle\Rector\FunctionLike\FunctionLikeToFirstClassCallableRector\Fixture;
+
+function (&$foo)
+{
+    return Foo::foo($foo);
+};
+
+?>
