+++ rules-tests/CodingStyle/Rector/FunctionLike/FunctionLikeToFirstClassCallableRector/Fixture/skip_functions_without_calls.php.inc
@@ -1,0 +1,12 @@
+<?php
+
+namespace Rector\Tests\CodingStyle\Rector\FunctionLike\FunctionLikeToFirstClassCallableRector\Fixture;
+
+function ($foo)
+{
+    return $foo;
+};
+
+fn ($foo) => $foo;
+
+?>
