+++ rules-tests/CodingStyle/Rector/FunctionLike/FunctionLikeToFirstClassCallableRector/Fixture/supports_varadic_and_unpack.php.inc
@@ -1,0 +1,22 @@
+<?php
+
+namespace Rector\Tests\CodingStyle\Rector\FunctionLike\FunctionLikeToFirstClassCallableRector\Fixture;
+
+function (...$foo)
+{
+    return FooBar::foo(...$foo);
+};
+
+fn (...$foo) => FooBar::foo(...$foo);
+
+?>
+-----
+<?php
+
+namespace Rector\Tests\CodingStyle\Rector\FunctionLike\FunctionLikeToFirstClassCallableRector\Fixture;
+
+FooBar::foo(...);
+
+FooBar::foo(...);
+
+?>
