+++ tests/Issues/CountArrayLongToShort/Fixture/count_to_empty_array_compare.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Rector\Tests\Issues\CountArrayLongToShort\Fixture;
+
+final class CountToEmptyArrayCompare
+{
+    public function run()
+    {
+        $data = [];
+
+        if (count($data) === 0) {
+        }
+    }
+}
+
+?>
